requirements.txt
# Django
django==4.1

# Django REST Framework
djangorestframework==3.13

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==20.1.0

# Swagger
drf-yasg

# CORS headers
django-cors-headers==3.9.0

pytz==2020.1

# Django filter
django-filter=23.3